requirements.txt
Cython==0.29.28
falcon==3.0.1
mysql-connector-python==8.0.28
